Split GUI sample load actions
